Remove purelb as it had some issues assigning ipv6
purelb was having issues figuring out which was the primary interface
for ipv6 addresses and would assign it to the internal iface.
The only solution I could find: setting the default iface to the host
iface broke more then it fixed as not all of the nodes have the same
primary interface name.